+++ src/utils.cpp
@@ -69,7 +69,7 @@ double sum(const cv::Mat& m) {
 }
 
 #ifdef HAVE_OPENCV_CUDACODEC
-void add_key_frame(std::vector<Key_Frame*>& key_frames, int delta, int frame_num,
+void add_key_frame(std::vector<Key_Frame*>& key_frames, double delta, int frame_num,
 	cv::cuda::GpuMat first_frame, cv::cuda::GpuMat second_frame) {
 	cv::Mat first_frame_cpu, second_frame_cpu;
 	if (!first_frame.empty()) {
@@ -82,7 +82,7 @@ void add_key_frame(std::vector<Key_Frame*>& key_frames, int delta, int frame_num
 }
 #endif
 
-void add_key_frame(std::vector<Key_Frame*>& key_frames, int delta, int frame_num,
+void add_key_frame(std::vector<Key_Frame*>& key_frames, double delta, int frame_num,
 	cv::Mat first_frame, cv::Mat second_frame) {
 	Key_Frame* new_key_frame = new Key_Frame;
 	new_key_frame->delta = delta;
